refactor: restructured forked scripts
Made fork of preprocessing tools from VRTool repository and restructured to the intended repository structure
