+++ README.md
@@ -1,0 +1,22 @@
+# Veiligheidsrendement Preprocessing #
+
+This is a repository with tools to translate results from the safety assessment and additional data to input for the VRTool
+
+## What is this repository for?
+
+* Quick summary
+* Version
+
+## How do I get set up? ##
+
+
+## Contribution guidelines ##
+
+* Writing tests
+* Code review
+* Other guidelines
+
+## Who do I talk to? ##
+
+* Repo owner or admin
+* Other community or team contact
