--- unit_costs.csv
@@ -1,13 +1,0 @@
-Description,Cost,Unit
-Inward starting costs,185, €/m
-Inward added volume,44,€/m3
-Outward added volume,36,€/m3
-Outward reused volume,14,€/m3
-Outward reuse factor,0.7,-
-Outward compensation factor,0.5,-
-Outward removed volume,19,€/m3
-Road renewal,670,€/m
-Sheetpile,340,€/m2
-Diaphragm wall,18700,€/m
-Vertical Geotextile,1330,€/m
-House removal,500000,€/unit
